* master: (371 commits)
  Shut up, my lovely compiler; I do not like your warnings
  Márcio Almada 	remove dead tokens: T_CHARACTER, T_BAD_CHARACTER.
  fix detection of mbstate_t with clang
  Add missing NEWS and UPGRADING for bug #68938
  moved the part of the test into an appropriate place
  Removed unused REGISTER_PDO_CONST_LONG
  Fix another invalid free of CG(interned_empty_string)
  add tests for #68996
  Fixed #68790 (Missing return)
  Fixed #68790 (Missing return)
  Update README.TESTING
  Small Spacing adjustments for better reading
  added ODBCVER to phpinfo()
  added PCRE JIT availability info to phpinfo()
  Improve tests for bug 67436
  Fixed issue #1041 (Fix build in OSX with --enable-dtrace)
  fix setting default ODBCVER value in config.w32
  updated NEWS
  updated NEWS
  Fixed bug #68964 Allowed memory size exhausted with odbc_exec
  ...

Conflicts:
	ext/session/session.c

+++ ext/hash/hash_md.c
@@ -135,7 +135,7 @@ PHP_NAMED_FUNCTION(php_if_md5)
 PHP_NAMED_FUNCTION(php_if_md5_file)
 {
 	char          *arg;
-	int           arg_len;
+	size_t        arg_len;
 	zend_bool raw_output = 0;
 	char          md5str[33];
 	unsigned char buf[1024];
